writerprofileform.tests refactored to accomodate Redux
